+++ src/EntityGraphQL/Compiler/EntityQuery/EntityQueryNodeVisitor.cs
@@ -124,7 +124,7 @@ namespace EntityGraphQL.Compiler.EntityQuery
                 return enumOrConstantValue;
             }
             var gqlField = schemaType.GetField(field, requestContext);
-            var exp = gqlField.GetExpression(gqlField.Resolve!, currentContext, null, null, new Dictionary<string, object>(), null, null, new List<GraphQLDirective>(), false);
+            var exp = gqlField.GetExpression(gqlField.ResolveExpression!, currentContext, null, null, new Dictionary<string, object>(), null, null, new List<GraphQLDirective>(), false);
             return exp!;
         }