+++ src/EntityGraphQL/Compiler/QueryWalkerHelper.cs
@@ -84,10 +84,10 @@ namespace EntityGraphQL.Compiler
                     throw new EntityGraphQLCompilerException($"Field {item.Name.Value} not found of type {schemaType.Name}");
                 var schemaField = (Field)schemaType.GetField(item.Name.Value, null);
 
-                if (schemaField.Resolve == null)
+                if (schemaField.ResolveExpression == null)
                     throw new EntityGraphQLCompilerException($"Field {item.Name.Value} on type {schemaType.Name} has no resolve expression");
 
-                var nameFromType = ((MemberExpression)schemaField.Resolve).Member.Name;
+                var nameFromType = ((MemberExpression)schemaField.ResolveExpression).Member.Name;
                 var prop = argType.GetProperty(nameFromType);
 
                 if (prop == null)