+++ src/EntityGraphQL/Schema/FieldExtensions/Filter/FilterExtension.cs
@@ -19,10 +19,10 @@ namespace EntityGraphQL.Schema.FieldExtensions
         /// <param name="field"></param>
         public override void Configure(ISchemaProvider schema, IField field)
         {
-            if (field.Resolve == null)
+            if (field.ResolveExpression == null)
                 throw new EntityGraphQLCompilerException($"FilterExtension requires a Resolve function set on the field");
 
-            if (!field.Resolve.Type.IsEnumerableOrArray())
+            if (!field.ResolveExpression.Type.IsEnumerableOrArray())
                 throw new ArgumentException($"Expression for field {field.Name} must be a collection to use FilterExtension. Found type {field.ReturnType.TypeDotnet}");
 
             listType = field.ReturnType.TypeDotnet.GetEnumerableOrArrayType()!;
@@ -31,7 +31,7 @@ namespace EntityGraphQL.Schema.FieldExtensions
             var args = Activator.CreateInstance(typeof(FilterArgs<>).MakeGenericType(listType));
             field.AddArguments(args);
 
-            isQueryable = typeof(IQueryable).IsAssignableFrom(field.Resolve.Type);
+            isQueryable = typeof(IQueryable).IsAssignableFrom(field.ResolveExpression.Type);
         }
 
         public override Expression GetExpression(Field field, Expression expression, ParameterExpression? argExpression, dynamic? arguments, Expression context, IGraphQLNode? parentNode, bool servicesPass, ParameterReplacer parameterReplacer)