+++ src/EntityGraphQL/Schema/IField.cs
@@ -45,7 +45,7 @@ namespace EntityGraphQL.Schema
         /// <param name="args"></param>
         /// <returns></returns>
         ExpressionResult? GetExpression(Expression fieldExpression, Expression? fieldContext, IGraphQLNode? parentNode, ParameterExpression? schemaContext, Dictionary<string, object> args, ParameterExpression? docParam, object? docVariables, IEnumerable<GraphQLDirective> directives, bool contextChanged);
-        Expression? Resolve { get; }
+        Expression? ResolveExpression { get; }
 
         IField UpdateExpression(Expression expression);